Revert "refactor: TSconfig 수정, auth, user, email 모듈 수정"
This reverts commit dd2821e07fb923f435e72425b8ca78e33aece2c3.

+++ src/apis/users/users.resolver.ts
@@ -4,9 +4,9 @@ import { UserService } from './users.service';
 
 import { UseGuards } from '@nestjs/common';
 import { GqlAuthAccessGuard } from 'src/commons/auth/gql-auth.guard';
-import { CurrentUser } from 'src/commons/auth/gql-user.param';
 import { CreateUserInput } from './dto/createUsers.input';
 import { UpdateUserInput } from './dto/updateUsers.input';
+import { CurrentUser } from 'src/commons/auth/gql-user.param';
 @Resolver('User')
 export class UserResolver {
   constructor(
@@ -17,7 +17,7 @@ export class UserResolver {
     description: '회원가입',
   })
   async createUser(@Args('createUserInput') createUserInput: CreateUserInput) {
-    return await this.userService.create(createUserInput);
+    return await this.userService.create({ createUserInput });
   }
 
   @UseGuards(GqlAuthAccessGuard)
@@ -25,32 +25,35 @@ export class UserResolver {
     description: '유저정보 수정',
   })
   async updateUser(
-    @CurrentUser('user') user: User,
+    @CurrentUser() currentUser: any,
     @Args('updateUserInput') updateUserInput: UpdateUserInput,
   ) {
-    return await this.userService.updateUser(user, updateUserInput);
+    return await this.userService.updateUser({
+      user: currentUser,
+      updateUserInput,
+    });
   }
 
   @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Boolean, {
     description: '유저정보 삭제',
   })
-  deleteUser(@CurrentUser('user') user: User) {
-    return this.userService.deleteUser(user);
+  deleteUser(@CurrentUser() currentUser: any) {
+    return this.userService.deleteUser({ user: currentUser });
   }
 
   @Query(() => Boolean, {
     description: '중복 이메일 확인',
   })
   async checkEmail(@Args('email') email: string): Promise<boolean> {
-    return await this.userService.checkEmail(email);
+    return await this.userService.checkEmail({ email });
   }
 
   @Query(() => Boolean, {
     description: '닉네임 중복 확인',
   })
   async checkNickname(@Args('nickname') nickname: string): Promise<boolean> {
-    return await this.userService.checkNickname(nickname);
+    return await this.userService.checkNickname({ nickname });
   }
 
   @Mutation(() => User, {
@@ -60,6 +63,6 @@ export class UserResolver {
     @Args('userEmail') userEmail: string,
     @Args('password') password: string,
   ) {
-    return await this.userService.resetPassword(userEmail, password);
+    return await this.userService.resetPassword({ userEmail, password });
   }
 }
